Updated Strawberry Artillery Fire
Updated the Strawberry Artillery Fire to fire a projectile straight upwards, before descending to the target from off-screen to the ground. Targets the last known position of the enemy as it begins falling, so lands just behind the enemy. Intended to hit the straggling enemies behind, rather than the initial target.