Vulture Implemented + Bug Fix
Implemented the Vulture Enemy, an enemy that can revive itself once and, while reviving, is immune to damage.

BUG FIX: Fixed a leaking logic issue where, for one or two frames, the Vulture could be considered dead after it's revive when stuck between two states, causing the wave count to be off and messing with the total enemies on screen, including the win loss condition.