Vulture Implemention
The vulture is a slow-flying despicably evil enemy with the power to rise from the dead. Upon it's first fatality, the Vulture will be knocked down, unable to move but immune to damage. After a brief time, the Vulture will get up, fully restoring it's health alongside shedding it's immunity and continuing on it's journey. Provides a fantastic distraction for smaller, faster enemies to sneak through while absorbing blows from distracted turrets.