BUG PATCHED: Sticky Projectiles
Fixed the bug where projectiles would "flicker" at a target position if the target died, remaining as an active projectile body for a few seconds or until another enemy container touched them.

Modified the projectile script code to prioritise moving the projectile regardless of state, as both conditions being checked resulted in double traversal on mobile rendering, causing a "flickering" effect.

The bug may still exist, but it seems to not now (so far)